Generate metrics on query runtime
QueryCanceledError were handled as API errors leading to a traceback
being uploaded to Sentry. A query timeout is not an API bug or
error condition.
We now capture all query runtimes in a metric for prometheus.
The query statement is too long to be used as query name or label,
so we log a short hash instead.